Move extractcode API to extractcode #2233
Also move extractcli tests there
