+++ src/extractcode/NOTICE
@@ -1,0 +1,56 @@
+Software license
+================
+
+Copyright (c) 2017 nexB Inc. and others. All rights reserved.
+http://nexb.com and https://github.com/nexB/scancode-toolkit/
+The ScanCode software is licensed under the Apache License version 2.0.
+Data generated with ScanCode require an acknowledgment.
+ScanCode is a trademark of nexB Inc.
+
+You may not use this software except in compliance with the License.
+You may obtain a copy of the License at: http://apache.org/licenses/LICENSE-2.0
+Unless required by applicable law or agreed to in writing, software distributed
+under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+CONDITIONS OF ANY KIND, either express or implied. See the License for the
+specific language governing permissions and limitations under the License.
+
+When you publish or redistribute any data created with ScanCode or any ScanCode
+derivative work, you must accompany this data with the following acknowledgment:
+
+  Generated with ScanCode and provided on an "AS IS" BASIS, WITHOUT WARRANTIES
+  OR CONDITIONS OF ANY KIND, either express or implied. No content created from
+  ScanCode should be considered or used as legal advice. Consult an Attorney
+  for any legal advice.
+  ScanCode is a free software code scanning tool from nexB Inc. and others.
+  Visit https://github.com/nexB/scancode-toolkit/ for support and download.
+
+
+Third-party software licenses
+=============================
+
+ScanCode embeds third-party free and open source software packages under various
+licenses including copyleft licenses. Some of the third-party software packages
+are delivered as pre-built binaries. The origin and license of these packages is
+documented by .ABOUT files.
+
+The corresponding source code for pre-compiled third-party software is available
+for immediate download from the same release page where you obtained ScanCode at:
+https://github.com/nexB/scancode-toolkit/
+or https://github.com/nexB/scancode-thirdparty-src/
+
+You may also contact us to request the source code by email at info@nexb.com or
+by postal mail at:
+
+  nexB Inc., ScanCode open source code request
+  735 Industrial Road, Suite #101, 94070 San Carlos, CA, USA
+
+Please indicate in your communication the ScanCode version for which you are
+requesting source code.
+
+
+License for ScanCode datasets
+=============================
+
+ScanCode includes datasets (e.g. for license detection) that are dedicated
+to the Public Domain using the Creative Commons CC0 1.0 Universal (CC0 1.0)
+Public Domain Dedication: http://creativecommons.org/publicdomain/zero/1.0/
