+++ tests/extractcode/data/cli/extract_nuget/00000xxxxx-00000xxxx-00000xxxx-00000xxxx/00000xxxxx-00000xxxx-00000xxxx-00000xxxx/microsoft.aspnet.identity.entityframework.2.2.1.nupkg
