+++ tests/extractcode/data/cli/extract_relative_path/basic.zip
